fix: temperature f32 → f64，消除 LLM 请求 400（参数超 2 位小数）
f32 0.7 经 serde_json 升位 f64 后序列化为 0.699999988079071，
被限制 2 位小数的模型 API 拒绝；f64 round-trip 精确

+++ symbio/src/symbio_core/schemas/model/model_config.rs
@@ -15,7 +15,7 @@ pub struct ModelConfig {
     pub api_key: Option<String>,
     pub model: String,
     #[serde(default = "default_temperature")]
-    pub temperature: f32,
+    pub temperature: f64,
     pub max_tokens: Option<u32>,
     pub system_prompt: Option<String>,
     #[serde(default = "default_max_context_tokens")]
@@ -31,7 +31,7 @@ pub struct ModelConfig {
     pub reasoning: Option<ReasoningConfig>,
 }
 
-fn default_temperature() -> f32 {
+fn default_temperature() -> f64 {
     0.7
 }
 fn default_max_context_tokens() -> u32 {
